rename packages and files
