+++ src/cli/cli.go
@@ -2,8 +2,8 @@ package cli
 
 import (
 	"flag"
-	"semangit/src/gitRepoManager"
-	"semangit/src/versionAnalyzers"
+	"semangit/src/repository"
+	"semangit/src/versionanalyzers"
 )
 
 type cli struct {
@@ -11,14 +11,14 @@ type cli struct {
 	fromRevision                   string
 	toRevision                     string
 	versionAnalyzerName            string
-	versionAnalyzersArgumentValues map[string]*versionAnalyzers.ArgumentValues
+	versionAnalyzersArgumentValues map[string]*versionanalyzers.ArgumentValues
 }
 
 func RunNewCli() cli {
 	c := cli{
-		fromRevision:                   gitRepoManager.RevisionNone,
-		toRevision:                     gitRepoManager.RevisionNone,
-		versionAnalyzersArgumentValues: make(map[string]*versionAnalyzers.ArgumentValues),
+		fromRevision:                   repository.RevisionNone,
+		toRevision:                     repository.RevisionNone,
+		versionAnalyzersArgumentValues: make(map[string]*versionanalyzers.ArgumentValues),
 	}
 	c.parseFlags()
 
@@ -41,27 +41,27 @@ func (c *cli) GetVersionAnalyzerName() string {
 	return c.versionAnalyzerName
 }
 
-func (c *cli) GetCurrentVersionAnalyzerArgumentValues() *versionAnalyzers.ArgumentValues {
+func (c *cli) GetCurrentVersionAnalyzerArgumentValues() *versionanalyzers.ArgumentValues {
 	return c.versionAnalyzersArgumentValues[c.versionAnalyzerName]
 }
 
 func (c *cli) parseFlags() {
 	flag.StringVar(&c.repoDir, "d", ".", "Repo root path. Defaults to current directory.")
-	flag.StringVar(&c.fromRevision, "f", gitRepoManager.RevisionNone, "From revision. A git reference to get version from.")
-	flag.StringVar(&c.toRevision, "t", gitRepoManager.RevisionNone, "From revision. A git reference to get version from.")
-	flag.StringVar(&c.versionAnalyzerName, "r", versionAnalyzers.VersionAnalyzerNameHelm, "Version analyzer. Defaults to 'helm'.")
+	flag.StringVar(&c.fromRevision, "f", repository.RevisionNone, "From revision. A git reference to get version from.")
+	flag.StringVar(&c.toRevision, "t", repository.RevisionNone, "From revision. A git reference to get version from.")
+	flag.StringVar(&c.versionAnalyzerName, "r", versionanalyzers.VersionAnalyzerNameHelm, "Version analyzer. Defaults to 'helm'.")
 	c.defineVersionAnalyzersFlags()
 	flag.Parse()
 
-	if c.GetToRevision() == gitRepoManager.RevisionNone {
+	if c.GetToRevision() == repository.RevisionNone {
 		panic("Provide TO revision (-t) to compare the version according to it.")
 	}
 }
 
 func (c *cli) defineVersionAnalyzersFlags() {
-	for _, versionAnalyzer := range versionAnalyzers.GetAllAnalyzers() {
+	for _, versionAnalyzer := range versionanalyzers.GetAllAnalyzers() {
 		argNamePrefix := versionAnalyzer.GetName() + "-"
-		argumentValues := versionAnalyzers.NewArgumentValues()
+		argumentValues := versionanalyzers.NewArgumentValues()
 		for _, argumentDefinition := range versionAnalyzer.GetExtraArgumentDefinitions() {
 			argumentValue := flag.String(argNamePrefix+argumentDefinition.Name, argumentDefinition.DefaultValue, argumentDefinition.Description)
 			argumentValues[argumentDefinition.Name] = argumentValue
