+++ src/repository/manager.go
@@ -1,4 +1,4 @@
-package gitRepoManager
+package repository
 
 import (
 	"github.com/go-git/go-git/v5"
