+++ src/versionanalyzers/helm_version_analyzer.go
@@ -1,4 +1,4 @@
-package versionAnalyzers
+package versionanalyzers
 
 import (
 	"gopkg.in/yaml.v3"
